+++ shenyu-bootstrap/pom.xml
@@ -152,10 +152,10 @@
         </dependency>
         <!--shenyu jwt end-->
 
-        <!--shenyu auth plugin start-->
+        <!--shenyu casdoor plugin start-->
         <dependency>
             <groupId>org.apache.shenyu</groupId>
-            <artifactId>shenyu-spring-boot-starter-plugin-auth</artifactId>
+            <artifactId>shenyu-spring-boot-starter-plugin-casdoor</artifactId>
             <version>${project.version}</version>
         </dependency>
         <!--shenyu auth plugin end-->
