+++ shenyu-spring-boot-starter/shenyu-spring-boot-starter-plugin/pom.xml
@@ -67,6 +67,6 @@
         <module>shenyu-spring-boot-starter-plugin-logging-pulsar</module>
         <module>shenyu-spring-boot-starter-plugin-logging-tencent-cls</module>
         <module>shenyu-spring-boot-starter-plugin-logging-clickhouse</module>
-        <module>shenyu-spring-boot-starter-plugin-auth</module>
+        <module>shenyu-spring-boot-starter-plugin-casdoor</module>
     </modules>
 </project>
